refactor: move html createElement stuff into own directory
