+++ src/html/create-element/attributes/index.js
